backend/requirements.txt
firebase-admin==6.5.0
pandas==2.2.1
python-dotenv==1.0.1
